+++ test/__scenarios__/zod-error.use-treeify-error.input.ts
@@ -20,6 +20,30 @@ export function format(error: z.ZodError) {
   return error.format();
 }
 
+// From Next.js tutorial https://nextjs.org/learn/dashboard-app/improving-accessibility
+const CreateInvoice = z.object({
+  customerId: z.string(),
+  amount: z.number(),
+  status: z.enum(["pending", "paid", "cancelled"]),
+});
+
+export async function createInvoice(formData: FormData) {
+  const validatedFields = CreateInvoice.safeParse({
+    customerId: formData.get("customerId"),
+    amount: formData.get("amount"),
+    status: formData.get("status"),
+  });
+
+  if (!validatedFields.success) {
+    return {
+      errors: validatedFields.error.flatten().fieldErrors,
+      message: "Missing Fields. Failed to Create Invoice.",
+    };
+  }
+
+  // ...
+}
+
 // Should not match
 export class ShouldNotTransform {
   format() {}